Insert link to CS 20.5
The CS20.5 website is now up. I'm the author of the CS20 and CS20.5 websites
and of course the CS20.5 website has a link back to CS21.
However, it's a little confusing the claim on the CS20.5 website that
CS 21 will happen in 2022, while the CS 21 webite itself still has a date for
2021. I thought it's the simpliest if I do the updates for you -
all you have to do is to merge this pull request.